+++ sdk/spring/azure-spring-boot-starter-active-directory/CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## 3.6.0-beta.1 (Unreleased)
 ### Breaking Changes
+- Deprecate aad.group.enable-full-list, use aad.group.allowed-group-ids=all instead.
 
 ### New Features
 - Support domain_hint in aad-starter.([#21517](https://github.com/Azure/azure-sdk-for-java/issues/21517))
-- Deprecate aad.group.enable-full-list, use aad.group.allowed-group-ids=all instead.[#21983](https://github.com/Azure/azure-sdk-for-java/issues/21983)
+- Use aad.group.allowed-group-ids=all means to configure all group ids, don't affect work for allowed-group-names.[#21983](https://github.com/Azure/azure-sdk-for-java/issues/21983)
 
 ## 3.5.0 (2021-05-24)
 ### New Features